Refactor getSelectors to selector in AxisScale
