+++ packages/graph/src/index.ts
@@ -10,4 +10,4 @@ export * from './themes';
 export * from './tooltip/Tooltip';
 export * from './tooltip/TooltipItem';
 export * from './utils/getAxisScale';
-export * from './utils/getRecordFieldSelectors';
+export * from './utils/getRecordFieldSelector';
